// src/app/models/user.ts
export class User {

    UserId: string = "00000000-0000-0000-0000-000000000000";
    FirstName: string = "";
    LastName: string = "";
    EmailAddress: string = "";
    Password: string = "";
    MobileCode: string = "+";
    Mobile: string  = "";
    UserRole: Number = 0;
    Client: Number = 0;
    Status: string = "";
}